Add definitions for extra keys in national layouts
The new files Keyboard_{de_DE,es_ES,fr_FR,it_IT}.h add macro definitions
for keys that could not be otherwise easily accessed with the library,
such as "e with grave accent" in the French and Italian layouts.

Only keys that produce either a non-ASCII character or a dead accent
when used unmodified are listed. For instance, the "pound sign" is not
listed in the Italian layout because, without the Shift modifier, that
key prints a "3". The macros are named after the character produced by
the unmodified key. For example, they key for the inverted exclamation
and question marks in the Spanish layout is named
`KEY_INVERTED_EXCLAMATION'.

This mimics what has already been done for the Danish and Swedish
layouts.